Move logging code over to verbosity.c
