+++ verbosity.h
@@ -148,6 +148,16 @@ void RARCH_ERR(const char *fmt, ...);
 #define RARCH_ERR_V RARCH_LOG_V
 #endif /* HAVE_LOGGER */
 
+void rarch_log_file_init(
+      bool log_to_file,
+      bool log_to_file_timestamp,
+      const char *log_dir);
+
+void rarch_log_file_deinit(void);
+
+void rarch_log_file_set_override(const char *path);
+
+
 RETRO_END_DECLS
 
 #endif
